requirements.txt
pytorch-fid
transformers
urllib3
torch
pillow
openai
ipython
